requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydub==0.25.1
librosa==0.10.1
numpy==1.26.4
scikit-learn==1.6.1
python-dotenv==1.0.0
soundfile==0.12.1
requests==2.31.0
pytest==8.0.0
joblib==1.4.0
scipy==1.13.1
setuptools
openai-whisper==20231117
keras


